useful paramter calculation with help of fits

Generates useful values for external use, with the help of parameterVector's of the fits.

Usage

parameterCalculation(parameterVector, stepSize = 1e-05)

Arguments

parameterVector

Output of multiple fit function sicegar::multipleFitFunction() that gives the variables related with sigmoidal or double sigmoidal fit.

stepSize

Step size used by the fitting algorithm. Smaller numbers gave more accurate results than larger numbers, and larger numbers gave the results faster than small numbers. The default value is 0.00001.

Value

Returns the expanded parameter vector. This vector includes useful derived values such as time and intensity of the start point, in addition to the standard values that the fit algorithms produce that are necessary to define the curves.

Examples

time <- seq(1, 24, 0.1)

#simulate intensity data with noise
noise_parameter <- 0.2
intensity_noise <- stats::runif(n = length(time), min = 0, max = 1) * noise_parameter
intensity <- sicegar::doublesigmoidalFitFormula(time,
                                            finalAsymptoteIntensityRatio = .3,
                                            maximum = 4,
                                            slope1Param = 1,
                                            midPoint1Param = 7,
                                            slope2Param = 1,
                                            midPointDistanceParam = 8)
intensity <- intensity+intensity_noise

dataInput <- data.frame(intensity = intensity, time = time)
normalizedInput <- sicegar::normalizeData(dataInput)
parameterVector <- sicegar::multipleFitFunction(dataInput = normalizedInput,
                                            dataInputName = "sample01",
                                            model = "doublesigmoidal",
                                            n_runs_min = 20,
                                            n_runs_max = 500,
                                            showDetails = FALSE)

if(parameterVector$isThisaFit){
       parameterVector <- sicegar::parameterCalculation(parameterVector)
}

print(t(parameterVector))
#>                                          [,1]             
#> finalAsymptoteIntensityRatio_N_Estimate  "0.3038072"      
#> finalAsymptoteIntensityRatio_Std_Error   "0.002237478"    
#> finalAsymptoteIntensityRatio_t_value     "135.7811"       
#> finalAsymptoteIntensityRatio_Pr_t        "7.362399e-218"  
#> maximum_N_Estimate                       "0.9812152"      
#> maximum_Std_Error                        "0.002565215"    
#> maximum_t_value                          "382.5079"       
#> maximum_Pr_t                             "1.503323e-318"  
#> slope1Param_N_Estimate                   "23.6693"        
#> slope1Param_Std_Error                    "0.3017369"      
#> slope1Param_t_value                      "78.4435"        
#> slope1Param_Pr_t                         "2.107121e-165"  
#> midPoint1Param_N_Estimate                "0.2914295"      
#> midPoint1Param_Std_Error                 "0.0007028794"   
#> midPoint1Param_t_value                   "414.6224"       
#> midPoint1Param_Pr_t                      "0"              
#> slope2Param_N_Estimate                   "23.73166"       
#> slope2Param_Std_Error                    "0.4621653"      
#> slope2Param_t_value                      "51.34885"       
#> slope2Param_Pr_t                         "3.113324e-126"  
#> midPointDistanceParam_N_Estimate         "0.3322909"      
#> midPointDistanceParam_Std_Error          "0.001429118"    
#> midPointDistanceParam_t_value            "232.5147"       
#> midPointDistanceParam_Pr_t               "4.926404e-270"  
#> residual_Sum_of_Squares                  "0.04576409"     
#> log_likelihood                           "657.056"        
#> AIC_value                                "-1300.112"      
#> BIC_value                                "-1276.015"      
#> isThisaFit                               "TRUE"           
#> startVector.finalAsymptoteIntensityRatio "0.9046085"      
#> startVector.maximum                      "0.6700916"      
#> startVector.slope1Param                  "125.3329"       
#> startVector.midPoint1Param               "0.1428656"      
#> startVector.slope2Param                  "14.94412"       
#> startVector.midPointDistanceParam        "0.5320635"      
#> dataScalingParameters.timeRange          "24"             
#> dataScalingParameters.intensityMin       "0.06192063"     
#> dataScalingParameters.intensityMax       "4.193228"       
#> dataScalingParameters.intensityRange     "4.131308"       
#> model                                    "doublesigmoidal"
#> additionalParameters                     "TRUE"           
#> finalAsymptoteIntensityRatio_Estimate    "0.3038072"      
#> maximum_Estimate                         "4.115623"       
#> slope1Param_Estimate                     "0.9862208"      
#> midPoint1Param_Estimate                  "6.994309"       
#> slope2Param_Estimate                     "0.988819"       
#> midPointDistanceParam_Estimate           "7.974981"       
#> dataInputName                            "sample01"       
#> betterFit                                "4"              
#> correctFit                               "20"             
#> totalFit                                 "41"             
#> maximum_x                                "10.98569"       
#> maximum_y                                "4.115623"       
#> midPoint1_x                              "6.918146"       
#> midPoint1_y                              "2.057811"       
#> midPoint2_x                              "15.04524"       
#> midPoint2_y                              "2.682989"       
#> slope1                                   "1.05211"        
#> slope2                                   "-0.7343884"     
#> finalAsymptoteIntensity                  "1.250356"       
#> incrementTime                            "3.911779"       
#> startPoint_x                             "4.962257"       
#> startPoint_y                             "0"              
#> reachMaximum_x                           "8.874036"       
#> reachMaximum_y                           "4.115623"       
#> decrementTime                            "3.901569"       
#> startDeclinePoint_x                      "13.09445"       
#> startDeclinePoint_y                      "4.115623"       
#> endDeclinePoint_x                        "16.99602"       
#> endDeclinePoint_y                        "1.250356"